requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
audioop-lts==0.2.2
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@4f352ca0a7450598833a770314f73e1f74e05f30
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
einops==0.8.1
faiss-cpu==1.13.0
fastapi==0.123.5
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.61.0
fsspec==2025.9.0
ftfy==6.3.1
google-auth==2.43.0
google-genai==1.52.0
gradio==6.0.2
gradio_client==2.0.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.12
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
roma==1.5.4
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.16.3
semantic-version==2.10.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
transformers==4.57.3
trimesh==4.10.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.234
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
